feat: add dummy implementation of IndexClientApi to core and http-client
