+++ packages/core/src/indexing/database-index-api.ts
@@ -1,0 +1,31 @@
+import type { StreamID } from '@ceramicnetwork/streamid'
+import type { BaseQuery, Pagination, Page } from '@ceramicnetwork/common'
+
+export interface IndexStreamArgs {
+  readonly streamID: StreamID
+  readonly model: StreamID
+  readonly controller: string
+  readonly lastAnchor: Date | null
+}
+
+/**
+ * Interface for an index backend.
+ */
+export interface DatabaseIndexApi {
+  /**
+   * This method inserts the stream if it is not present in the index, or updates
+   * the 'content' if the stream already exists in the index.
+   * @param args
+   */
+  indexStream(args: IndexStreamArgs): Promise<void>
+
+  /**
+   * Query the index
+   */
+  page(query: BaseQuery & Pagination): Promise<Page<StreamID>>
+
+  /**
+   * Initialize connection to a database.
+   */
+  init(): Promise<void>
+}
