+++ packages/core/src/indexing/not-implemented-error.ts
@@ -1,3 +1,6 @@
+/**
+ * Helps to track what we have to implement in the near future.
+ */
 export class NotImplementedError extends Error {
   constructor(name: string) {
     super(`NOT IMPLEMENTED: ${name}`)
