+++ packages/core/src/indexing/sqlite/__tests__/sqlite-index-api.test.ts
@@ -4,7 +4,7 @@ import { DataSource } from 'typeorm'
 import { asTimestamp, SqliteIndexApi } from '../sqlite-index-api.js'
 import { StreamID } from '@ceramicnetwork/streamid'
 import { listMidTables } from '../init-tables.js'
-import { IndexStreamArgs } from '../../types.js'
+import { IndexStreamArgs } from '../../database-index-api.js'
 import csv from 'csv-parser'
 import * as fs from 'fs'
 import knex, { Knex } from 'knex'
